refactor: RHI Vulkan
Put files in folder
